gui/requirements.txt
gymnasium
pygame
pillow
imageio
imageio-ffmpeg
